+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - automatically register for all alarm types
   - create Docker image
   - create Debian package
+### Fixed
+- API
+  - support messages with DispatcherBody data longer than 256 bytes
 
 ## [0.0.3] - 2022-12-19
 ### Changed